chore(workers): rebucket yaniv casino->solo for WASM size headroom
The casino worker sits at the 1MB gzipped free-tier limit (~1,048,585 bytes,
a few bytes over on high-variance TinyGo builds), causing intermittent
tinygo-build failures and risking the master->deploy build. Move Yaniv (a
self-contained Israeli draw-and-discard game, ~50KB source across its files)
from the casino worker to solo (which had ~110KB headroom): frees casino's
binary while keeping solo well under the limit.

To make Yaniv movable, extract the generic `sliceOrEmpty[T]` helper out of the
casino-only TexasHoldemBonus.go into a new untagged internal/domain file so it
compiles into every worker (Yaniv used it; previously it was reachable only in
the casino binary).

Follows the documented rebucket procedure: retag Yaniv's 10 production files
casino->solo, flip its registry Category, move its RegisterKVGame call from
casino/ to solo/, bump registry_test worker counts (casino 65->64, solo
50->51), repoint gameApi.ts, and update docs/cloudflare-workers.md. Verified
all four worker WASM binaries (casino/classic/solo/extra) compile. No game
logic changed.
