+++ docs/cloudflare-workers.md
@@ -4,10 +4,10 @@ Games are deployed to Cloudflare Workers as WASM binaries via TinyGo. **Four** w
 
 | Worker | Entry point | Games |
 |--------|-------------|-------|
-| **casino** | `cmd/workers/casino/main.go` | Table & poker games (blackjack, baccarat, poker, holdem, omaha, omahahilo, bigo, bigohilo, shortdeck, pineapple, crazypineapple, irishpoker, indianpoker, videopoker, deuceswild, jokerpoker, threecard, fourcardpoker, caribbeanstud, texasholdembonus, ultimatetexasholdem, mississippistud, sevencardstud, paigow, chinesepoker, letitride, reddog, razz, badugi, deucetoseven, spanish21, casinowar, dragontiger, blackjackswitch, oasispoker, russianpoker, casinoholdem, highcardflush, yaniv, tressette, bourre, napoleon, mighty, bridge, skat, belote, tarneeb, sheepshead, doppelkopf, mus, tute, sueca, fortyfives, twentynine, courtpiece, ecarte, threecardbrag, teenpatti, kemps, pishti, cuarenta, fivecardstud, faro, openfacechinese) |
+| **casino** | `cmd/workers/casino/main.go` | Table & poker games (blackjack, baccarat, poker, holdem, omaha, omahahilo, bigo, bigohilo, shortdeck, pineapple, crazypineapple, irishpoker, indianpoker, videopoker, deuceswild, jokerpoker, threecard, fourcardpoker, caribbeanstud, texasholdembonus, ultimatetexasholdem, mississippistud, sevencardstud, paigow, chinesepoker, letitride, reddog, razz, badugi, deucetoseven, spanish21, casinowar, dragontiger, blackjackswitch, oasispoker, russianpoker, casinoholdem, highcardflush, tressette, bourre, napoleon, mighty, bridge, skat, belote, tarneeb, sheepshead, doppelkopf, mus, tute, sueca, fortyfives, twentynine, courtpiece, ecarte, threecardbrag, teenpatti, kemps, pishti, cuarenta, fivecardstud, faro, openfacechinese) |
 | **classic** | `cmd/workers/classic/main.go` | Trick-taking, matching & fishing (hearts, spades, pitch, twotenjack, callbreak, briscola, oldmaid, doubt, daifugo, bigtwo, sevens, crazyeights, ohhell, speed, gofish, pinochle, pigtail, durak, war, fiftyone, whist, pageone, trash, president, cassino, spiteandmalice, shithead, nertz, slapjack, egyptianratscrew, tonk, sixcardgolf, truco, klaverjas, manille, marias, sedma, solowhist, knockoutwhist, nap, preference, spoilfive, doudizhu, tichu, scopa, scopone, escoba, bezique, cuckoo, spoons, labellelucie, simplesimon, doubleklondike) |
 | **extra** | `cmd/workers/extra/main.go` | Overflow size bucket — games rebalanced off the other three workers (see ADR-0032). Rummy family moved off `solo` (ginrummy, conquian, chinchon, threethirteen, canasta, handandfoot, contractrummy, kalooki, burraco, rummy500); plus solitaires streetsandalleys, kingalbert, flowergarden, fortyandeight, agnes, sultan; plus the trick-takers jass, gaigel, tysiac, calabresella, ombre, king, cinch, and loo; plus the fishing games basra and tablanet; plus the banking game trenteetquarante; plus the poker-vying pot games guts, bouillotte, primero, and the Pass-the-Trash poker game anaconda; plus the stops chip-betting game michigan; plus the raise/bluff trick-taker watten; plus the contract-rummy game carioca; plus the three-deck Canasta variant samba; plus the 13-card Indian Rummy (indianrummy); plus the Italian shared-table Rummy Machiavelli (machiavelli); plus the multi-deck Mexican Rummy Panguingue (pan); plus the Hungarian contract trick-taker Ulti (ulti); plus the first non-52 procedural-render-path games (ADR-0033): the Wizard trick-taker (wizard), the kabufuda banking game Oicho-Kabu (oichokabu), the 57-card Rook (rook), the hanafuda games Koi-Koi (koikoi), Go-Stop (gostop) and Hachi-Hachi (hachihachi), the 78-card French Tarot (frenchtarot), and the Austrian tarock game Königrufen (koenigrufen). |
-| **solo** | `cmd/workers/solo/main.go` | Solitaire & rummy (klondike, freecell, seahaventowers, cruel, spider, spiderette, pyramid, tripeaks, memory, cribbage, golf, clocksolitaire, fortythieves, canfield, yukon, russiansolitaire, scorpion, wasp, accordion, pokersquares, montecarlo, calculation, bakersdozen, beleagueredcastle, sevenbridge, crescent, gaps, eightoff, penguin, acesup, barbu, macau, mao, thirtyone, tienlen, osmosis, fivehundred, schnapsen, gongzhu, bristol, bidwhist, easthaven, bakersgame, euchre, piquet, russianbank, blackhole; plus — bucketed here for size headroom once the extra worker hit the 1 MB limit — the 78-card Italian tarot trick-taker Scarto (scarto, ADR-0033), the 54-card Baden tarock game Cego (cego), and the 54-card Chinese climbing game Zheng Shangyou (zheng)) |
+| **solo** | `cmd/workers/solo/main.go` | Solitaire & rummy (klondike, freecell, seahaventowers, cruel, spider, spiderette, pyramid, tripeaks, memory, cribbage, golf, clocksolitaire, fortythieves, canfield, yukon, russiansolitaire, scorpion, wasp, accordion, pokersquares, montecarlo, calculation, bakersdozen, beleagueredcastle, sevenbridge, crescent, gaps, eightoff, penguin, acesup, barbu, macau, mao, thirtyone, tienlen, osmosis, fivehundred, schnapsen, gongzhu, bristol, bidwhist, easthaven, bakersgame, euchre, piquet, russianbank, blackhole; plus — bucketed here for size headroom once the extra worker hit the 1 MB limit — the 78-card Italian tarot trick-taker Scarto (scarto, ADR-0033), the 54-card Baden tarock game Cego (cego), and the 54-card Chinese climbing game Zheng Shangyou (zheng); plus the Israeli draw-and-discard game Yaniv (yaniv), rebucketed off `casino` for size headroom) |
 
 The worker entry points (`cmd/workers/{casino,classic,solo}/main.go`) are thin shells that blank-import the matching `internal/infrastructure/games/<category>` sub-package and call `games.RegisterCategory(mux, games.Category…)`. The registry itself (`internal/infrastructure/games/registry.go`) stores `{Name, Category}` for each game; the human-readable descriptions live in a separate `descriptions.go` map (build-tagged `//go:build !js || !wasm` to keep them out of the WASM binaries). The Web-server factories live in `games_server.go` (excluded from WASM via build tags) and the Worker bindings live in per-category sub-packages — this split is what keeps each Cloudflare Worker binary under the 1 MB gzipped free-tier limit by letting TinyGo dead-code-eliminate the games from the other two categories.
